requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
mysqlclient
Pillow
